requirements.txt
pygame>=2.5.0
pymunk>=6.6.0
numpy>=1.26.0
pyrr>=0.10.3
tree-sitter>=0.21.0
tree-sitter-c-sharp>=0.21.0
jinja2>=3.1.0
pyyaml>=6.0
pytest>=8.0.0
pyright>=1.1.350
ruff>=0.5.0
